Feat/batch collation (#4)
* reorg `data` module

* move `samplers`
