+++ mol_gnn/data/datasets/molecule.py
@@ -6,10 +6,11 @@ from rdkit import Chem
 from sklearn.preprocessing import StandardScaler
 from torch.utils.data import Dataset
 
+from mol_gnn.data.models.datum import Datum
 from mol_gnn.types import Mol
-from mol_gnn.data.mixins import _DatapointMixin, _MolGraphDatasetMixin, Datum
 from mol_gnn.featurizers import VectorFeaturizer, GraphFeaturizer, MolGraphFeaturizer
 from mol_gnn.utils.chem import make_mol
+from mol_gnn.data.datasets.mixins import _DatapointMixin, _MolGraphDatasetMixin
 
 
 @dataclass
