+++ mol_gnn/models/mpnn.py
@@ -9,7 +9,7 @@ from torch import nn, Tensor, optim
 
 from mol_gnn.data import MpnnBatch
 from mol_gnn.data.models.graph import BatchedGraph
-from mol_gnn.data.batch import MpnnBatch
+from mol_gnn.data.models.batch import MpnnBatch
 from mol_gnn.data.models.graph import BatchedGraph
 from mol_gnn.nn import Aggregation, MessagePassing, Predictor, LossFunction
 from mol_gnn.nn.encoder import GraphEncoder
